agent/requirements.txt
openai>=1.40.0
requests>=2.32.0
feedparser>=6.0.11
apscheduler>=3.10.4
PyMuPDF>=1.24.0
openpyxl>=3.1.0
pyyaml>=6.0
